requirements.txt
pandas==0.20.3
numpy==1.13.3
pypng==0.0.18
gunicorn==19.8.1
flask==1.0.2